semantic_model_generation/requirements.txt
loguru==0.7.3
pandas==2.2.3
protobuf==3.20.3
pydantic==2.10.4
python-dotenv==1.0.1
ruamel.base==1.0.0
snowflake_connector_python==3.12.3
snowflake_connector_python==3.12.4
